refactor: removed class HintList and changed Hint into an interface (#41)
